+++ crates/rhdl-fpga/src/serial_bus/mod.rs
@@ -27,6 +27,7 @@ pub mod mfm_encoder;
 pub mod midi;
 pub mod mipi_dbi_type_b;
 pub mod mipi_dbi_type_c;
+pub mod nand_flash_async;
 pub mod one_wire_master;
 pub mod rs485_master;
 pub mod sent_rx;